requirements.txt
djangorestframework==3.15.2
django-filter==24.3
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
python-decouple==3.8
psycopg2-binary==2.9.10
